create a script to prepare an ubuntu system for midipatch

+++ prep_ubuntu.sh
@@ -9,5 +9,7 @@ apt-get update
 # install dependencies
 apt-get install git-core build-essential g++-7 zlib1g-dev libasound2-dev liblilv-dev npm libsqlite3-dev
 npm install -g inliner
-ln -fs /usr/bin/nodejs /usr/bin/node
 
+if [ ! -f /usr/bin/node ]; then
+  ln -fs /usr/bin/nodejs /usr/bin/node
+fi
